requirements.txt
mysql_connector_repackaged==0.3.1
pygame==2.6.1
